% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gen_segs}
\alias{gen_segs}
\title{Function to generate a number of sequence segments of a particular length from an input sequence}
\usage{
gen_segs(seq, len = 90, num = 1000)
}
\arguments{
\item{seq}{Sequence to draw segments from. Can be a character vector or \code{\link{DNAString}} object of length 1}

\item{len}{Length of the segments to be generated}

\item{num}{Number of segment to generate}
}
\value{
A \code{\link{DNAStringSet}} object containing the generated sequence segmenta
}
\description{
Function to generate a number of sequence segments of a particular length from an input sequence
}

